+++ src/constants/settings_constants.js
@@ -1,0 +1,5 @@
+export const SETTINGS_KEY = 'bm_settings';
+export const INIT = 'settings_init';
+export const SAVE = 'settings_saved';
+
+export const CHANGE = 'settings_changed';