+++ src/constants/sync_constants.js
@@ -1,6 +1,1 @@
-const PROTOCOL = 'http://';
-const DOMAIN = '127.0.0.1';
-const PORT = '3000';
-export const SERVER_URL = `${PROTOCOL}${DOMAIN}:${PORT}`;
-
 export const DB_NAME = 'BookmarksManager';