+++ src/helpers/construct_server_url.js
@@ -1,0 +1,7 @@
+export default function constructServerUrl(url, port) {
+  if (typeof port === 'number') {
+    return `${url}:${port}`;
+  } else {
+    return url;
+  }
+}