+++ CHANGELOG.md
@@ -2,3 +2,12 @@
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+### Added
+
+* Form to add server url and port for connection status check and synchronization
+
+### Changed
+
+* Pass server url from localStorage to the synchronization code and connection check
+* Automatically close the panel if the add/search operations succeed